Move context by arch
