+++ src/ExecBlock/X86/CMakeLists.txt
@@ -1,0 +1,4 @@
+# Add QBDI target
+set(SOURCES "${CMAKE_CURRENT_LIST_DIR}/Context_X86.cpp")
+
+target_sources(QBDI_src INTERFACE "${SOURCES}")
