+++ src/ExecBlock/context_switch/windows-X86_64_Stub.asm
